add ktx2 textures
